requirements.txt
fastapi
pydantic
mysql-connector-python
mysql
SQLAlchemy==1.4.39
python-dotenv
uvicorn==0.18.2
PyMySQL==1.0.2
line-bot-sdk==2.4.2